requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
attrs==25.3.0
audioop-lts==0.2.1
azure-ai-inference==1.0.0b9
azure-core==1.35.0
certifi==2025.8.3
charset-normalizer==3.4.2
cowsay==5.0
cowsay-py==0.1.2
discord.py==2.5.2
frozenlist==1.7.0
idna==3.10
isodate==0.7.2
multidict==6.6.3
propcache==0.3.2
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
typing_extensions==4.14.1
urllib3==2.5.0
yarl==1.20.1
